+++ src/clojure/parkour/io/dux.clj
@@ -5,11 +5,12 @@
                      (mapreduce :as mr)]
             [parkour.mapreduce (sink :as snk)]
             [parkour.io (dseq :as dseq) (dsink :as dsink) (mux :as mux)]
-            [parkour.util :refer [returning]])
+            [parkour.util :refer [returning prev-reset!]])
   (:import [clojure.lang IFn]
            [org.apache.hadoop.conf Configurable]
            [org.apache.hadoop.mapreduce Job TaskInputOutputContext]
            [org.apache.hadoop.mapreduce OutputFormat RecordWriter Counter]
+           [org.apache.hadoop.mapreduce TaskAttemptContext]
            [parkour.hadoop Dux$OutputFormat]))
 
 (def ^:private ^:const confs-key
@@ -126,68 +127,98 @@ format only) file basename `base`."
   ([context oname base key val]
      (-> context (get-sink oname base) (snk/emit-keyval key val))))
 
-(defn named-keyvals
+(defn ^:private close-rws
+  [^TaskAttemptContext context]
+  (let [[jobs _ rws] (dux-state context)
+        taid (.getTaskAttemptID context)
+        rws (prev-reset! rws nil)]
+    (doseq [[name rws'] rws, :let [tac (-> jobs (get name) (mr/tac taid))]
+            rw (->> rws' vals (map (comp w/unwrap deref)))]
+      (.close ^RecordWriter rw ^TaskAttemptContext tac))))
+
+(defmacro ^:private with-close-rws
+  [context & body]
+  `(try ~@body (finally (close-rws ~context))))
+
+(defn map-output
+  "Sink as (reducer-bound) base map output, as `mr/sink-as` kind `kind`."
+  [kind]
+  (let [f (snk/sink-fn* kind)]
+    (fn [context coll]
+      (with-close-rws context
+        (f context coll)))))
+
+(defn combine-output
+  "Sink as (reducer-bound) base combiner output, as `mr/sink-as` kind `kind`."
+  [kind] kind)
+
+(defn ^:private named
+  "Base function for `named-`* functions."
+  ([f oname]
+     (fn [context coll]
+       (with-close-rws context
+         (let [sink (get-sink context oname)]
+           (reduce (fn [_ t]
+                     (f sink t))
+                   nil coll)))))
+  ([f context coll]
+     (with-close-rws context
+       (reduce (fn [_ [oname k v :as x]]
+                 (let [t (if (= 2 (count x)) k [k v])
+                       sink (get-sink context oname)]
+                   (f sink t)))
+               nil coll))))
+
+(def ^{:arglists '([oname] [context coll])}
+  named-keyvals
   "Sink as key-val pairs to named output, with name provided as `oname` or as
 first element of three-element tuples."
-  ([oname]
-     (fn [context coll]
-       (reduce (fn [_ [key val]]
-                 (write context oname key val))
-               nil coll)))
-  ([context coll]
-     (reduce (fn [_ [oname key val]]
-               (write context oname key val))
-             nil coll)))
+  (partial named snk/emit-keyval))
 
-(defn named-keys
+(def ^{:arglists '([oname] [context coll])}
+  named-keys
   "Sink as keys to named output, with name provided as `oname` or as first
 element of two-element tuples."
-  ([oname]
-     (fn [context coll]
-       (reduce (fn [_ key]
-                 (write context oname key nil))
-               nil coll)))
-  ([context coll]
-     (reduce (fn [_ [oname key]]
-               (write context oname key nil))
-             nil coll)))
+  (partial named snk/emit-key))
 
-(defn named-vals
+(def ^{:arglists '([oname] [context coll])}
+  named-vals
   "Sink as values to named output, with name provided as `oname` or as first
 element of two-element tuples."
-  ([oname]
-     (fn [context coll]
-       (reduce (fn [_ val]
-                 (write context oname nil val))
-               nil coll)))
-  ([context coll]
-     (reduce (fn [_ [oname val]]
-               (write context oname nil val))
-             nil coll)))
+  (partial named snk/emit-val))
 
-(defn prefix-keyvals
+(defn ^:private prefix
+  "Base function for `prefix-`* functions."
+  ([f oname]
+     (fn [context coll]
+       (with-close-rws context
+         (reduce (fn [_ [base k v :as x]]
+                   (let [t (if (= 2 (count x)) k [k v])
+                         sink (get-sink context oname base)]
+                     (f sink t)))
+                 nil coll))))
+  ([f context coll]
+     (with-close-rws context
+       (reduce (fn [_ [oname base k v :as x]]
+                 (let [t (if (= 3 (count x)) k [k v])
+                       sink (get-sink context oname base)]
+                   (f sink t)))
+               nil coll))))
+
+(def ^{:arglists '([oname])}
+  prefix-keyvals
   "Sink as key-val pairs to named output `oname`, with file prefix as first
 element of three-element tuples."
-  [oname]
-  (fn [context coll]
-    (reduce (fn [_ [base key val]]
-              (write context oname base key val))
-            nil coll)))
+  (partial prefix snk/emit-keyval))
 
-(defn prefix-keys
+(def ^{:arglists '([oname])}
+  prefix-keys
   "Sink as keys to named output `oname`, with file prefix as first element of
 two-element tuples."
-  [oname]
-  (fn [context coll]
-    (reduce (fn [_ [base key]]
-              (write context oname base key nil))
-            nil coll)))
+  (partial prefix snk/emit-key))
 
-(defn prefix-vals
+(def ^{:arglists '([oname])}
+  prefix-vals
   "Sink as values to named output `oname`, with file prefix as first element of
 two-element tuples."
-  [oname]
-  (fn [context coll]
-    (reduce (fn [_ [base val]]
-              (write context oname base nil val))
-            nil coll)))
+  (partial prefix snk/emit-val))